+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Priorities can be added, renamed, recolored, and reordered in settings
+
+### Fixed
+
+- The import dialog offered the built-in statuses and priorities instead of the configured ones
+
 ## [1.8.0] - 2026-07-03
 
 ### Added